renamed components, pages, routers to components_temp, pages_temp. routers_temp
