+++ src/components_temp/hotels/HotelCard/HotelCard.jsx
