+++ src/components_temp/hotels/HotelCard/HotelCard.styles.jsx
