+++ src/components_temp/hotels/HotelMapView.jsx
