+++ src/components_temp/hotels/HotelRating.jsx
