+++ src/components_temp/hotels/ListHotels.jsx
