+++ src/components_temp/hotels/PhotoGallery.jsx
