+++ src/components_temp/hotels/RatingBadge.jsx
