+++ src/pages_temp/HotelDetailsPage.jsx
