+++ src/pages_temp/HotelsPage.jsx
